Initial Upload
This script will handle everything that has to do with the cursor, including hover and click events and well as changing textures. The start function sets the cursor image to the default texture decided by the developer. The OnCursorEnterWater function changes the cursor's texture to the assigned texture while hovering over an appropriate water button while OnCursorExitWater returns the cursor's texture back to its default upon not hovering over the same button.